+++ src/selector/misc/tokens.ts
@@ -5,4 +5,4 @@ import {DateTimeSelectorOptions} from './datetimeSelector.interface';
 /**
  * Injection token used for obtaining datetime selector configuration
  */
-export const DATE_TIME_SELECTOR_CONFIGURATION: InjectionToken<DateTimeSelectorOptions<any>> = new InjectionToken<DateTimeSelectorOptions<any>>('DATE_TIME_SELECTOR_CONFIGURATION');
+export const DATE_TIME_SELECTOR_CONFIGURATION: InjectionToken<Partial<DateTimeSelectorOptions>> = new InjectionToken<Partial<DateTimeSelectorOptions>>('DATE_TIME_SELECTOR_CONFIGURATION');